Ambient Study Tests/pyproject.toml
[tool.poetry]
name = "ambient study tests"
version = "0.1.0"
description = ""
authors = ["Bruno Alves Comitre <bruno-comitre@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
PySimpleGUI = "^4.60.1"
httpx = "^0.23.0"
SQLAlchemy = "^1.4.37"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
